Changed post processing toggle to be an alias
Having reports of PP randomly turning off with no way to repro, I decided to remove the auto-PP-reset thing and just change the UI toggle to just toggle PP as if you did it in graphics options, that means it gets saved to config and all that.
Also a various tweaks to tooltips for PP and particle quality.